ML Project/ML Project/requirements.txt
streamlit
transformers
numpy
torch
openai-whisper
